Solvd Y2016Day10.
First get the input bucket values assigned to my custom robot classes. Then run all the comparisons (failing and keeping the comparison in the queue of comparisons to run if the robot hasnt got the values yet), until the comparisons list is empty.

This completes both part one and two, outputs kept an integer array.